Fix a long-standing bug in RapidBuild64/003-interp/03-mod_customize.sh that was incorrectly removing tkinter

+++ RapidBuild64/003-interp/03-mod_customize.sh
@@ -9,9 +9,10 @@ if [ "${INCMARIADB}" != "Y" ]; then
   rm -rf ${DIR}/usr/${LIBDIR}/perl5/vendor_perl/auto/DBD/mysql
 fi
 
-# WRAP THIS WITH INCTCL INCTK or something
-rm -f ${DIR}/usr/${LIBDIR}/python?.?/lib-dynload/_tkinter.so
-rm -f ${DIR}/usr/${LIBDIR}/python?.?/lib-dynload/_tkinter.cpython-*.so
+if [ "${INCTCL}" != "Y" ]; then
+  rm -f ${DIR}/usr/${LIBDIR}/python?.?/lib-dynload/_tkinter.so
+  rm -f ${DIR}/usr/${LIBDIR}/python?.?/lib-dynload/_tkinter.cpython-*.so
+fi
 
 if [ -x ${DIR}/usr/bin/gtkless.py ]; then
   ln -s gtkless.py ${DIR}/usr/bin/gtkless
